adding the rest of the props
