+++ lib/_shield-advanced-stack.ts
@@ -4,7 +4,7 @@ import { Prerequisites } from "./types/config";
 import { RuntimeProperties } from "./types/runtimeprops";
 import { aws_fms as fms } from "aws-cdk-lib";
 import { ManagedServiceData, SubVariables } from "./types/fms";
-
+import { ShieldConfig } from "./types/config";
 // export interface StackProps extends cdk.StackProps {
 //   readonly prerequisites: Prerequisites;
 //   runtimeProperties: RuntimeProperties;
@@ -28,9 +28,9 @@ import { ManagedServiceData, SubVariables } from "./types/fms";
 
 // export interface shield_props extends cdk.StackProps {
 //   accounts: ShieldAccount[];
-//  
+//
 // }
-// 
+//
 // const shieldProps: shield_props = {
 //   accounts: [
 //     {
@@ -50,49 +50,53 @@ import { ManagedServiceData, SubVariables } from "./types/fms";
 //   cfnpolicy: cfnShieldPolicyProps
 // };
 
+
+
+// export interface shield_props extends cdk.StackProps {
+//   remediationEnabled: boolean | cdk.IResolvable;
+//   resourceTypeList?: Array<string>;
+//   resourceType: string;
+//   policyName: string;
+//   includeMap?: fms.CfnPolicy.IEMapProperty | cdk.IResolvable;
+//   excludeMap?: fms.CfnPolicy.IEMapProperty | cdk.IResolvable;
+//   excludeResourceTags: boolean | cdk.IResolvable;
+//   securityServicePolicyData:
+//     | cdk.IResolvable
+//     | fms.CfnPolicy.SecurityServicePolicyDataProperty;
+// }
 export interface shield_props extends cdk.StackProps {
-  resourceType: string;
+  readonly config: ShieldConfig;
 }
 export class ShieldStack extends cdk.Stack {
-  constructor(scope: Construct, id: string, props: shield_props ) {
+  constructor(scope: Construct, id: string, props: shield_props) {
     super(scope, id, props);
+    // const cfnShieldPolicyProps: fms.CfnPolicyProps = {
+    //   // remediationEnabled - should be true
+    //   remediationEnabled: props.remediationEnabled,
+    //   // https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_fms.CfnPolicy.html#resourcetype:~:text=fms%2Dpolicy%2Dresourcetags-,resourceType,-%3F
+    //   resourceTypeList:props.resourceTypeList,
+    //   resourceType: props.resourceType,
+    //   policyName: props.policyName,
+    //   includeMap: props.includeMap,
+    //   excludeMap: props.excludeMap,
+    //   excludeResourceTags: props.excludeResourceTags,
+    //   //https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_fms.CfnPolicy.html#:~:text=Specification%20for%20SHIELD_ADVANCED%20for%20Amazon%20CloudFront%20distributions
+    //   securityServicePolicyData: props.securityServicePolicyData,
+    // };
 
-    // might be unnecessary
-    const preProcessRuleGroups: never[] = [];
-    const postProcessRuleGroups: never[] = [];
-    let loggingConfiguration;
-    const managedServiceData: ManagedServiceData = {
-      type: "SHIELD_ADVANCED",
-      defaultAction: { type: "ALLOW" },
-      preProcessRuleGroups: preProcessRuleGroups,
-      postProcessRuleGroups: postProcessRuleGroups,
-      overrideCustomerWebACLAssociation: false,
-      loggingConfiguration: {
-        logDestinationConfigs: [loggingConfiguration || ""],
-      },
-    };
     const cfnShieldPolicyProps: fms.CfnPolicyProps = {
       // remediationEnabled - should be true
-      remediationEnabled: true,
+      remediationEnabled: props.config.remediationEnabled,
       // https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_fms.CfnPolicy.html#resourcetype:~:text=fms%2Dpolicy%2Dresourcetags-,resourceType,-%3F
-      resourceTypeList: [
-        "AWS::CloudFront::Distribution",
-        "AWS::ElasticLoadBalancingV2::LoadBalancer",
-        "AWS::ElasticLoadBalancing::LoadBalancer",
-        "AWS::EC2::EIP",
-      ],
-      resourceType: props.resourceType,
-      policyName: "ShieldAdvancedPolicyCdkTest",
-      includeMap: { account: ["962355891833"] },
-      excludeMap: {},
-      excludeResourceTags: false,
+      resourceTypeList:props.config.resourceTypeList,
+      resourceType: props.config.resourceType,
+      policyName: props.config.policyName,
+      includeMap: props.config.includeMap,
+      excludeMap: props.config.excludeMap,
+      excludeResourceTags: props.config.excludeResourceTags,
       //https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_fms.CfnPolicy.html#:~:text=Specification%20for%20SHIELD_ADVANCED%20for%20Amazon%20CloudFront%20distributions
-      securityServicePolicyData: {
-        type: "SHIELD_ADVANCED",
-        managedServiceData: cdk.Fn.sub(JSON.stringify(managedServiceData), {}),
-      },
+      securityServicePolicyData: props.config.securityServicePolicyData,
     };
-
     const fmspolicy = new fms.CfnPolicy(
       this,
       "CfnPolicy",
